+++ frontend/src/components/common/NavBar.tsx
@@ -1,25 +1,19 @@
 import { useNavigate } from "react-router-dom";
 const NavBar:React.FC = () => {
     const navigate = useNavigate();
-    const handleGoLogin = () =>{
-        navigate("/login");
-    }
-    const handleGoSignup = () =>{
-        navigate("/signup");
-    }
     const handleGoAdmin = () =>{
         navigate("/admin");
     }
+    const handleGoShop = () =>{
+        navigate("/");
+    }
     return (
         <div>
-            <div onClick={handleGoLogin}>
-                로그인
-            </div>
-            <div onClick={handleGoSignup}>
-                회원가입
-            </div>
             <div onClick={handleGoAdmin}>
                 어드민페이지
+            </div>
+            <div onClick={handleGoShop}>
+                샵페이지
             </div>
         </div>
     );
